+++ direcs/src/obstacleCheckThread.h
@@ -198,8 +198,8 @@ class ObstacleCheckThread : public QThread
 		/**
 		This are areas to be ignored when checking the laser distances.
 		*/
-		static const short int AREA1 = 1;
-		static const short int AREA2 = 2;
+		static const int AREA1 = 1;
+		static const int AREA2 = 2;
 
 		// Every thread sleeps some time, for having a bit more time fo the other threads!
 		// Time in milliseconds
